WIP Splitting core functionality into modules
